src/main/java/com/jfuerste/recipewebapp/converters/IngredientToIngredientCommand.java
package com.jfuerste.recipewebapp.converters;

import com.jfuerste.recipewebapp.commands.IngredientCommand;
import com.jfuerste.recipewebapp.domain.Ingredient;
import lombok.Synchronized;
import org.springframework.core.convert.converter.Converter;
import org.springframework.lang.Nullable;
import org.springframework.stereotype.Component;

@Component
public class IngredientToIngredientCommand implements Converter<Ingredient, IngredientCommand> {

    private final UnitOfMeasureToUnitOfMeasureCommand uomConverter;

    public IngredientToIngredientCommand(UnitOfMeasureToUnitOfMeasureCommand uomConverter) {
        this.uomConverter = uomConverter;
    }

    @Synchronized
    @Nullable
    @Override
    public IngredientCommand convert(Ingredient source) {
        if (source == null){
            return null;
        }

        IngredientCommand command = IngredientCommand.builder()
                .amount(source.getAmount())
                .id(source.getId())
                .description(source.getDescription())
                .uom(uomConverter.convert(source.getUom()))
                .build();

        return command;

    }

}
